requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
google-generativeai

streamlit
requests

selenium
webdriver-manager

beautifulsoup4
lxml

python-dotenv
